+++ benchmarker/modules/problems/gemm/data.py
@@ -1,2 +1,4 @@
 def get_data(params):
+    if params["framework"] == "torch":
+        print("custom torch tensor")
     return "dummy data"
